V44: deep-debt evolution — linalg extraction, typed errors, docs sweep
- Extract tridiag eigensolver into shared `linalg` module (transport re-exports)
- Add `error` module with typed `InputError` (LengthMismatch, InsufficientData, OutOfRange)
- Evolve 5 APIs from assert! to Result: jackknife, block_jackknife, finite_size_extrapolate, chi_squared, percentile
- Add capability-based UID discovery (XDG/procfs, zero unsafe)
- Enrich derives: GridFitConfig +Debug/Clone/Copy, EighError +Clone/PartialEq/Eq
- Idiomatic cast: prng::next_u64 as u64 → u64::from()
- V44 ToadStool handoff with full absorption guidance and delegation inventory
- Archive V43, update wateringHole/README with V44 as active
- Sweep stale delegation counts (32→39, 9→7 pending) across ~20 docs
- Fix three_tier_parity dispatch assertion to match 39 active + 7 pending
- Update CHANGELOG, CONTROL_EXPERIMENT_STATUS, README, CONTRIBUTING
- Update whitePaper (STUDY, METHODOLOGY, baseCamp, experiments, CROSS_SPRING_EVOLUTION)
- Update specs (README, BARRACUDA_REQUIREMENTS, PAPER_REVIEW_QUEUE)
- Update wateringHole/CROSS_SPRING_SHADER_EVOLUTION
- All tests pass, zero clippy warnings, zero unsafe

Made-with: Cursor

+++ control/bistable_switching/bistable_switching.py
@@ -50,7 +50,7 @@ def hill(x: float, k: float, n: float) -> float:
     if x <= 0:
         return 0.0
     xn = x ** n
-    return xn / (k ** n + xn)
+    return float(xn / (k ** n + xn))
 
 
 # ---------------------------------------------------------------------------
