+++ control/multisignal_qs/multisignal_qs.py
@@ -51,14 +51,14 @@ def hill(x: float, k: float, n: float) -> float:
     if x <= 0:
         return 0.0
     xn = x ** n
-    return xn / (k ** n + xn)
+    return float(xn / (k ** n + xn))
 
 
 def hill_repress(x: float, k: float, n: float) -> float:
     if x <= 0:
         return 1.0
     kn = k ** n
-    return kn / (kn + x ** n)
+    return float(kn / (kn + x ** n))
 
 
 # ---------------------------------------------------------------------------
